requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.44.1
sentence-transformers==2.7.0
